Issue #11163: Enforce file size on Java inputs for InputConstructorsDeclarationGrouping
